Make sql.GeneratorSuite and move the existing testcases.
